app theme: use 'background', not 'surfaceVariant'
I thought that surfaceVariant would be the appropriate color to tweak and use as the background.
But there is a color from the RN Paper theme called 'background' which is more appropriate.

I noticed that 'surfaceVariant' is used for DataTable, and I think some Dialogs, so changing it for the background has some side effects.
But 'background', as far as I can tell, only affects the background of the LabelScreen. I tweaked the color in the process and also noted where surfaceVariant is used.