fix activities on draft trips by setting timezone
Error was thrown in AddNoteButton -> getPrefillTimes because unprocessed/draft trips did not have `timezone` defined in their local_dt objects.
In services.js we compute the properties for unprocessed trips in `points2TripProps`. To get the dt objects we call `moment2localdate`, which takes a tz argument, but we did not use it.

We have access to the timezone names from `startPoint` and `endPoint` metadata, so let's use them.
Now activities can be added to draft trips, and the prefilled timestamps are correct.

Tested trips from various different timezones just to be sure it was not reverting to my local timezone.