replace non-clickable <IconButton>s with <Icon>s
Icons that aren't clickable should be marked as presentational / not interactive and not focusable so that screen readers don't pick up on them.
Up to this point we've been using IconButton for everything, clickable or not. Let's make a new component Icon, which is a wrapper around IconButton that marks it as not interactive and disables the padding so it's just the plain icon.

From now on, clickable icons will use <IconButton> and plain, presentational icons will use <Icon>.