aria: add labels, roles to DiaryCard components
In TripCard, the left and right panels were reordered. They still appear on the left and right respectively, but now the right panel is ordered before the left panel.
This is so that screenreaders read the right panel (which has the basic trip info like date, duration, locations) before the left panel. This should be clearer to blind/visually impaired users who rely on audio cues.

Visually, the panels are still rendered the same as before because the flexDirection of their parent is now 'row-reverse'