fix extra pipelineRange request every time labeled
Resolves a bug tracked in https://github.com/e-mission/e-mission-docs/issues/948

The pipeline range should only be fetched when the label screen is initialized or refreshed.
When a trip is labeled, we should not need to fetch the pipeline range - this was happening because the old `getUnprocessedLabels` function called `getPipelineRangeTs` every time it was called.

This function was rewritten into `getUnprocessedInputs` in `timelineHelper.ts`. It will accept the pipelineRange as an argument.

Now in LabelTab, only on init / refresh will we request the pipelineRange.
Subsequent calls to `getUnprocessedLabels` will use the already-fetched pipelineRange.

We will also handle errors better - if retrieving pipeline range fails (network error), we should show error popup and stop loading spinner.

Testing done:
-Labeled several trips - all labels appeared within ~1 second.
-Disconnected network, tried labeling trips - all labels appeared within ~1 second.
-While network disconnected, tried refreshing label screen - popup "Error while loading pipeline range". Same result on app launch while network disconnected.

extra tidying note:
- removed getAngularService("DiaryHelper") from LabelTab because it no longer used (the functions that were previously used there had already been moved to the new Typescript file diaryHelper.ts)