on network error, don't show TimelineScrollList
The last commit handles network errors better - we show a popup error if the pipeline range can't be fetched.
In that situation, we don't want to render anything at all for the TimelineScrollList. `listEntries` is null so there's no cards to show, and we shouldn't show the header/footer ("Load more") buttons either. Let's modify this condition to only show the FlashList if `listEntries` is defined.