finish implementation of EnketoModal
add the validateAndSave function to EnketoModal, along with all the logic needed to support it.
Functions have been rewritten from enketo/service.js and enketo/launch.js. I did not rewrite enketo/answer.js, so the EnketoSurveyAnswer service is still used in enketoHelper.ts.
Nothing has dramatically changed here - I just rewrote the same logic in Typescript.

One thing that has changed is the way that validation errors appear. If something is wrong in the survey response, we used to show an IonicPopup describing the issue. IonicPopup has a lower z-index than the React Native Modal. Rather than try to hack the z-index and compete for the foreground, I opted to show these popups using `window.alert` instead.
I tested it on both Android and iOS, and found that the window alerts look like the native popups on their respective OS. It actually provide a much nicer UX for errors and I will move the Logger over to using this as well.