add caching for loading surveys
This handy function can be used for any URL resource that we'd like to be cached to localStorage.
For right now we'll just use this for any surveys loaded by EnketoModal.

I decided to put this in a new file called commHelper.ts. This is a new function, but the existing Angular CommHelper service will be rewritten at some point and those new functions can go in this file too.