use Dialogs for multilabel popups
Instead of using Ionic popups and separate templates for the mode and purpose select menus, we can utilize a Dialog containing Radio buttons using components from React Native Paper.

This is much better in terms of accessibility - the Ionic popups were not able to be navigated to by keyboard / screen reader controls, while the Dialog gains focus when opened because it is wrapped in a Modal.

The UX for inputting 'other' has been changed as well. Instead of having 'other' entry be a separate popup, I incorporated it into the same Dialog.
When the "Other" radio button is chosen, a text entry and Save button will appear right below it allowing the user to specify their custom label.