EnketoModal: tweak "Dismiss" and "Save" buttons
The button to dismiss the Enketo Modal has been incorporated into the Enketo header so it doesn't take up a whole extra row.
The accessibility roles / visibility have been tweaked for both the dismiss and save buttons so that they are exposed correctly to the accessibility tree as clickable buttons without extra nesting.