refactor SurveyOptions to TypeScript
SurveyOptions is just an object with some definitions for the MULTILABEL vs ENKETO configurations. We don't need an Angular service for this; let's just make it an exported const object and import it where needed.
Also with type safety!

In TripCard, I noticed that we check the config and set surveyOpt as state.
We already do this in LabelTab and can just use the surveyOpt from there via useContext. This must have been there from before we set up the LabelTabContext to share state. No reason to do this in both places.

The directive 'linkedsurvey' is not used anymore so it is being removed.