EnketoModal: receive click events on <button>s
Typically to receive clicks in React Native you need a Touchable or Pressable element, or something that composes one, and then use the onPress prop as a handler.

But for Enketo, (web-based/HTML) we aren't working with React Native code; these are HTML elements.
So I thought I had to wrap buttons in Pressables to receive any click events.
It turns out that for HTML <button>s, we can use onClick instead of onPress and it works the same. So we don't need those extra <Pressable>s, and the <button> already gives itself an aria role of button.

Add a little bit of styling to make the <button>s behave the same as the <Pressable>s did, and we're set.

While here, I cleaned up the HTML and added a comment describing the template as well.