EnketoModal: make survey take up full height
if the survey does not have enough content on a page to fill the entire height of the screen, we should have it grow to fill the remaining space. This way, the footer is always at the bottom.
Adding flex: 1 to everything in the modal up until the enketo-plugin realizes this behavior