create new confirmHelper.ts
This Typescript file will replace the ConfirmHelper service, and also support reading multilabel options from a URL specified in the dynamic config, instead of from separate JSON files (https://github.com/e-mission/e-mission-docs/issues/945). But if label_options is not in the dynamic config, we just fall back to the old method.

This file is type safe :)