use confirmHelper instead of Angular ConfirmHelper
We created confirmHelper.ts which was a rewrite of the Angular service ConfirmHelper.
There are a couple differences in how they are used. Because ConfirmHelper was an Angular service, it was a singleton object that contained a set of defined properties.
confirmHelper.ts is a module so I chose to keep these as module-wide variables. And instead of exposing these directly it has lazy getter functions. If the getter functions have been called before it's returning the module-wide saved result. Otherwise it gets the needed value.
So this approach shouldn't result in a performance drop; it's just a different way of implementing a singleton pattern.

The new confirmHelper doesn't have `isProgram` and `mode_studied` since we can just grab those straight from the appConfig directly in the one place they are used.
For everything else, the rewritten confirmHelper has a replacement / the equivalent of what was called from the old ConfirmHelper.