adjust DiaryButton UI
DiaryButton was the first component written in React and it was a little bit off in terms of appearance.
We can slim down the padding so more text can fit, and make the icon, if present, spaced evenly from the text.

We don't need to pass the text as a separate prop; we can accept the content text as the child and embed it that way.

Let's also type the Props as an extension of ButtonProps, since this component is basically a wrapper around the RN Paper <Button>.