try locally hosted configs on __DEV__, & refactor
In the devapp or on dev builds, we have the __DEV__ global set true. We can utilize this to simplify the dev process of testing different configs.
Instead of manually changing the downloadURL in dynamic_config.js to localhost:9090 when we want to test other configs, we can have dynamic_config.js automatically check localhost:9090 for locally served configs. If it doesn't find a local config, it fallbacks to downloading from github.

On production, where __DEV__ is false, it will not check for local configs and will go straight to github (same behavior as before).

In the process, I refactored this to use `fetch` instead of Angular's `$http` because it will make it easier to rewrite this file later.