implement useDerivedProperties
The composite and confirmed trip/place objects we receive from the server don't include all of the strings we actually show in the UI. We do some formatting to these before showing these as 'display' props.
Instead of 'populating' trip/place objects by mutating these additional properties into them when they are loaded, let's make a memoized hook that extracts the derived properties for a given timelineEntry.

Then we can pull from these downstream where they're used. And we won't have to worry about 'populating' trips/place before they are used for display - the information contained in the trip/place objects should now be almost exactly the same as what is on the server. (Eventually unifying the type definitions to match 100% is a long-term goal)