refactor into StartEndLocations.tsx
All of the diary cards (TripCard, PlaceCard, UntrackedTimeCard), as well as the LabelDetailsScreen, all show the names of some locations, with only a few differences between.

A bunch of code was repeated across each of these files, so we can refactor this into a reuasable component that we can call StartEndLocations.