on load config, cache resources referenced by URL
In our config file, we may reference external resources by URL in the form of Enketo surveys or a 'label options' JSON file.
To optimize this, we should have these be fetched + cached asynchronously when the config is first loaded. This way we won't have to wait for them to load when they are needed.