commit diaryTypes.ts [WIP]
With the goal of eventual type safety with regard to composite trip / confirmed place objects, I have been tracking down which properties are actually present on the server and which ones are computed on the phone.

Eventually, it would be great to have our type definitions for the Composite/Confirmed objects line up 1:1 between our JS and Python code.
However, there is still some work to do before we can apply these type definitions anywhere (the "populated trip" properties still need to be reworked)