fix dynamic config loading in dev environments
In dev environments (where __DEV__ == true), we will check localhost:9090 for locally hosted configs. if we are not serving at that moment, we should just skip.
The fetch throws an error if localhost:9090 is not active - we should catch this, skip the locally hosted config and check fetch from github instead.