fix: add explicit annotations to logger.ts
We use a Babel plugin `angularjs-annotate` to fix our bad, implicitly annotated, Angular code. This way, it still works when bundled. But we only have Babel apply this to .js files.
I moved a little bit of Angular code to .ts (the "Logger" service). I don't expect we will usually put Angular code in .ts files, so I'm just going to explicitly type it in the one place we do.