add typings for metric data
This typing is based on the data we receive from the server.
For each metric, we receive an array containing "days" of metric data. Each day has a few props (ts, fmt_time, nUsers, local_dt) plus any number of properties with the `label_` prefix that represent the measurment for a particular label.

Having this be type-safe from the start should make it much easier to realized the desired visualizations!