reformat MetricsCard to work with new MetricsTab
MetricsCard will now accept 'metricDataDays' as a prop and derive the chart data from this.
We can employ `useMemo` for this so that iff metricDataDays changes, chartData updates.

Prior to this, MetricsCard had an IconButton on the right side allowing users toggle between 'details' and 'graph' view.
To be clearer to the user, I have changed this to a SegmentedButtons. This way, users see both their options side by side and can select the desired one.

A few colors were added in appTheme.ts - surfaceDisabled and onSurfaceDisabled.
These are colors that already exist in the React Native Paper default theme, but we are overriding for our own purposes.
In the SegmentedButtons on MetricsCard, the surfaceDisabled color is used for the button that is not checked