fill out MetricDetails
With the new MetricsTab, MetricsDetails will receive 'metricDataDays', sum up the totals for each label, and display these in a 2-column.
The display is the same as the old dashboard, but now the logic to sum up the values is here in this component instead of jumbled in with everything else.