adjust padding in MetricsCard
The chart does not need so much padding - we want to be able to show as much detail as we can.
MetricDetails should show with 2 columns, but could use some padding in between.

So let's remove all the padding from the card content since its children will have padding anyway.