add serve-prod script to package.serve.json
It will be useful to test out 'production' JS (which is bundled + minified) on the devapp, so we can debug any issues related to bundling + minification. `npm run serve-prod` will accomplish that.
Since this script will use the 'prod' webpack configuration, we'll also have to add the angularjs-annotate plugin to devDependencies.