fix vertical scrolling on new Dashboard tab
I noticed that when the content was tall enough to allow vertical scrolling, you could not scroll to see the bottom even though the content was in a ScrollView.
Upon inspection, the new Dashboard tab was growing infinitely tall, past the bottom of its Angular parent container.

We've had this happen a couple times before when embedding React components into Angular views.

The way we get around this is with a CSS class called 'fill-container'. I modified it slightly to only affect the *direct* child, not all grand- and great-grand-children. Without this modification, the entire tab would be hidden from view.