show date ranges for weeks on ActiveMinutesCard
To make it clearer what data is being used to calculate week-wise metrics, "Past week" and "Previous week" should be annotated with the specific dates they encompass.
For this to not clutter the chart, I wanted these to be on a newline, like this:

Past Week
8/21-8/27

ChartJS doesn't support this by itself - `\n` is just rendered as a space - but we can remedy this by adding a callback function under ChartJS options > 'scales' > 'ticks' that splits a label into an array if it has `\n`.