readable labels on new dashboard
the dashboard works with the label keys "e.g. 'walk', 'drove_alone'. To show these as readable text we can repurpose a function from `confirmHelper.ts` which is used for "other" text.
Let's rename it to be more generic, make it work with more than just 1 underscore, and export it so we can use it on the dashboard too,