diaryHelper: rename 'percentages' to detectedModes
We use a field called 'percentages' to display the mode icon and percentage on the trip card above the map.
However, this is an object containing more than just the percentage - it has mode, icon, color, and percentage.

Soon, we plan to not even show the percentage in unimodal trips. So before that, let's rename this to something that is more descriptive.