create <ModesIndicator>
I am extracting this out to a separate component since its logic is getting more complex as we show either the detected modes or the labeled mode on the map.

This new component lists the detected or labeled modes at the top of the leaflet map on a trip card.
If the trip has not been labeled, it will show "Detected:" with icon(s) for the mode(s) during the trip.
If the trip has been labeled, it will show the labeled mode (which is in trip.userInput.MODE) and use the icon and color of that mode's `baseMode` (which is defined in labelOptions config).
(e.g. 'free_shuttle' will show a purple bus icon if its base mode is "BUS")