add baseModes to label options
When a trip is labeled, we need to know what icon and color to show above the map. To get this information, we will require each labelOption to have a baseMode, allowing us to map it onto a set of recognized modes ('BaseModes' in diaryHelper.ts)

A couple new BaseModes were added, since the Material Design Icons library has icons some of the more niche modes that we probably didn't have icons for previously.