refactor getBaseMode functions
We have 2 ways by which we lookup a base mode - one is with the key and the other is when a trip has been labeled, we look up its `baseMode` from the labelOptions.

Let's refactor to keep these in the same place in diaryHelper.ts and name them so it is clearer what they do.