update map trajectory colors when trip is labeled
We reworked ModeIndicator so that once a trip is labeled, it show the labeled mode icon and color.
Now we need to have the map update the color of the trajectory to match the mode icon.

services.js > confirmedPoints2Geojson is where the geojson is created and the color is specified there.
If the trip is labeled, we will call getBaseModeOfLabeledTrip. For this, we need the labelOptions object, so we need to pass this down through the call stack from LabelTab, down to compositeTrip2Geojson, and finally to confirmedPoints2Geojson.

Then we can get the labeled base mode, or if the trip isn't labeled, use the sensed_mode as we did before.

To get the map to actually update in real time, we need to have the geojson recomputed in repopulateTimelineEntry, and we also need LeafletView to listen for changes to 'geojson'.

Once that is all addressed, we can finally see the trajectory colors change in real time upon changing the labeled mode.