fix labeled mode not showing if detected=UNKNOWN
We were returning early if detectedModes was empty or the only detected mode was UNKNOWN.
But once that trip is labeled, we won't be using detectedModes, and we should still show the mode that was labeled. So let's move these checks down to and else if condition, after we've already checked for a user label.
If both if and else if conditions fail, then nothing will be rendered.