refactor compositeTrip2Geojson-> useGeojsonForTrip
If we want to be able to switch back and forth between 'detected modes' and 'labeled mode', we need to approach the generation of geojsons in a different way.
Precomputing and recomputing is not the best option because the map needs to change every time someone labels or toggles between 'detected'/'labeled'.

We'll refactor the functions that did this in the Timeline service and rewrite new versions in timelineHelper. The trip geojson will be computed by useGeojsonForTrip (which caches previously computed geojsons so performance does not take a hit). Then we can call this where it is needed instead of precomputing and storing 'geojson' in the trip object.