use 'useGeojsonForTrip', toggle labeled/detected
The geojson object is no longer stored in the trip; we will access it in the 2 places it's used by using the new 'useGeojsonForTrip' function.
On the details screen, if a trip has been labeled, we will be able to toggle between 'detected' and 'labeled' modes using SegmentedButtons (if set to 'detected', no 'labeledMode' will be passed to useGeojsonForTrip).
If the trip is not labeled, we cannot toggle and just show "Detected Modes", aiming to match the style