support labeled mode on TripSectionsDetails
if the toggle on the details screen is set to show the labeled mode, we want to be able to show that here instead of the sensed sections.
If the trip is labeled, we treat it as unimodal regardless of what it was detected as. So if we are showing the labeled mode, we can construct a 1-item array and use the properties of the whole trip for distance and duration.
And for the mode icon + color, we can call getBaseModeOfLabeledTrip, just as we have done elsewhere in the codebase.

Then fill these in place of the detected 'section' properties.

Some of the 'useDerivedProperties' variable names were changed here to make them more understandable to other maintainers in the context of the UI