rename OverallTripDescriptives, TripSectionsDescriptives
Sticking with the term "descriptives" to avoid confusion with "details".
(Since the entire screen is called "trip details", we don't want to have a smaller component also called "trip details"

And since the "TripDescriptives" component (describing the whole trip) is only shown if there are multiple sections showing (to avoid redundancy), it's clearer if we call it "OverallTripDescriptives".