make inferred labels yellow instead of orange
Inferred labels used to be a bright yellow with white text on them. Since white text on yellow is hard to read, the button color was shifted to a deeper orange to achieve  better legibility / color contrast ratio.
However, we received some user feedback that the deep orange felt intimidating. To the user, this seemed like some kind of warning or error.
They asked us to bring back the yellow color.

So I reworked DiaryButton to work with dark text and accept a custom border color.
Now, inferred labels show on a pale yellow button with dark text, framed by a yellow border.