allow gaps in daily active minutes line chart
It's actually important to push records for every day on this chart, even if the values are undefined or otherwise falsy.
For example, most people don't bike every day - so we need to show gaps in the line graph on the days they don't, and the only way to do that is to pass falsy values through.