embed <profile-settings> directly in main.js
We do not need main-control.html as an intermediate between main.js and ProfileSettings; we can include the <profile-settings> element directly as a 'template'. It still uses general-settings as a controller.

In ProfileSettings we would access the Angular settings scope by getting the child '.ion-view' of the main-control element. Now we have removed this and the settings scope lives on the main-control element itself.

Lastly, the fill-container class needed to be applied, just as was described in https://github.com/JGreenlee/e-mission-phone/commit/b6dfb4f9cc543720774eab1400c798ecf424474e when similar issues affected the Dashboard tab.

The result of all this is that there is no weird gap above ProfileSettings scrolling, and scrolling is smoother and more consistent.
-main-control.html is removed as no longer needed