refactor modeColors
chartData itself depends on modeColors, and nothing else uses modeColors. modeColors and chartData both depend on the same thing (chartDatasets) and should update whenever it does.
So we can just have modeColors be computed inside of the body of chartData.

And to ensure we have labelOptions defined, we'll add that to the dependency array, returning blank data if it isn't defined