fix warnings in MetricsCard
React warns us about the missing 'key' prop if we don't put a key whenever we iterate and repeat components, so a `key` was added to this View.

The way the brackets were arranged on line 89 was actually causing a single space " " to be rendered, which React was also warning us about in the console.