fix start_loc / end_loc missing on draft trips
https://github.com/e-mission/e-mission-docs/issues/971#issuecomment-1711113612
Before the migration, we filled in the coordinates of start_loc and end_loc by getting longitude and latitude and putting them as an array. When draft trips were added to the label screen, we changed to getting the `loc` property which contains the coordiantes array.

However, loc does not exist for locations that haven't been processed yet. So, we need to construct these from the properties we are guaranteed to have, like before.
Since it's a pretty small object, we can just do this inline.