trim whitespace around custom labels
https://github.com/e-mission/e-mission-docs/issues/969#issuecomment-1710702181
Whitespace at the beginning or end of a user-defined custom label could get us into trouble because we `split` by ' ', which could give blank strings at the beginning/end of the resulting array - when we go through each word to capitalize it we'd hit an error. So we must trim the string when we convert readable -> label key.
And In case there are any custom labels already stored in the DB that have leading/trailing whitespace, we'll also trim when doing label key -> readable