handle errors better
On the label screen, if there are errors we want to show user-facing popups so people can report the errors effectively and we can address them.

We will do this whenever trips are 1) loaded and 2) populated, since these areas were missing any kind of error catches

logger.ts was slightly modified to show a 'header' in the window.alert popup, with the title of the error between bars like:

━━━━
<Error title>
━━━━
<Error body...>

This is fairly primitive but I think it works for now; we can later explore "Alert" libraries from the community if we want to.