filter draft trips that are <100m and <5 min
https://github.com/e-mission/e-mission-docs/issues/966
We often mistakenly get 'stubs' of trips while people are staying at one place (just sitting at home, etc) - the sensing detects a transition but it is not a real trip.

We filter for these on the server and usually they don't make it through the pipeline, but we also want to filter them out for draft trips (while they exist on the phone/ before the server filters them)

There is already a filter step here to filter out any 'undefined' entries. Let's tack on some extra criteria so we also filter anything with distance <100m and duration <5 min.

Note:
this file is going to be rewritten before too long so I'm not going to clean anything else up while I'm here.