fix enketo trip survey responses on details screen
If a survey response is recorded from UserInputButton on the Label details screen, the button should reflect the newly recorded response.
This wasn't working because UserInputButton did not call repopulateTimelineEntry, allowing the entire Label tab context to update (which is the paradigm we are using for MultilabelButtonGroup and AddNoteButton.
Instead we were just writing into the timelineEntry object on the 'userInput' field (the old/Angular way of doing things - and React is not aware of this mutation happening)
Unifying this with the approach of AddNoteButton resolves the issue.

Also refactored UserInputButton in general: we do not need both displayLabel and isFilled as state.
If no response label exists, that's the same as isFilled=false
The responseLabel is derived from the properties within timelineEntry - so useMemo is more efficient than useState+useEffect here.

Lastly, let's ensure the log statements in onResponseSaved are using logger.ts, not plain console statements.