fix label details screen for draft trips
Unprocessed ('draft') trips don't have any sections but we still want to show descriptives for them.
We'll construct a single section to be displayed, based on the overall trip, in much the same way that we do when we're showing the labeled mode instead of the detected modes.

We don't know what mode draft trips are, so we won't have a base mode to use for icon+color - we can fallback to UNPROCESSED (which shows a question mark in grey).
Draft trips won't have 'text' either - it will stay undefined and not show up in the UI.

Once a draft trip is labeled, it will be treated the same as any other labeled trip.