big refactor of BarChart & LineChart
Instead of repeating a bunch of code between BarChart and LineChart, let's refactor this into a common Chart component that works with a type of either 'bar' or 'line'.

Then BarChart and LineCharts will simply be composed of a Chart with the appropriate type specified.
Any logic specific to BarCharts or LineCharts can be put in the respective files.

Currently, LineChart has no specific logic and the only thing BarCharts have specifically is the 'meter' prop.