make space for annotation at top of chart
If the annotations sit at the top of a horizontal bar chart, they will cover up much of the top bar.
We can offset them so they sit higher, and set clip=false so they can show above the chart area, but then they crowd the legend.

There isn't a good way to add padding/margin to a ChartJS legend, so we have this workaround in 'plugins'