fix dateRange issue
The use of UTC time here was inconsistent with how MetricsDateSelect works - MetricsDateSelect uses the local timezone for date selection.
DateTime.utc returns a DateTime object which may be on a different day than the day in the local timezone.

MetricsDateSelect works properly for this and the initial load should follow suit.