adjust carbon meter gradient
We will now use the same yellow-orange-red gradient for all bars that have a gradient, instead of making a specific gradient for each bar. This allows the gradients on each bar to line up.

The gradient will only apply to bars that exceed meter.low. Anything below low is green. Anything in middle is yellow, becoming orange-red as it gets higher. If it goes above high, it will start to become dark red up to high+20.

https://github.com/e-mission/e-mission-phone/pull/1018#issuecomment-1718083794