refactor: remove darkenForBorder + fix
We already have a function that serves this purpose, so let's just use it (darkenOrLighten)
It turns out that it was actually flipped from what was intended, though -- negative numbers should equate to darkening and positive to lightening.
It didn't really matter before because it was used for deduping where one duplicate is always lightened while the other is darkened.