fix segmentDaysByWeeks
if we had dates for the 1st through 15th (15 days), this function would mistakenly return the 8th through 14th and the 1st through 7th.

We actually want the 9th through 15th and the 2nd throgh 8th, because it should be relative to the most recent day.

To fix this we iterate over the days backwards, not forwards. With this approach we also don't have to reverse the weeks when we return them because they are already with the most recent first.