add App component; use BottomNavigation for tabs
Instead of using Ionic's tabbing implementation, we're going to use `BottomNavigation` from RN Paper, included in a central App component.

This App component will have an AppContext, which can include any app-wide variables/functions the tabs need to share, including the loaded dynamic config - this way we can just load it once and use it everywhere.

So App.tsx is replacing main.html, as well as the logic that was in MainCtrl.
The metrics tab is only shown in 'MULTILABEL' configurations (same as the implementation was in MainCtrl).

Onboarding hasn't been implemented yet; so there is a placeholder for it right now